add parser for packed-refs format
